DapsDriver: Update return value to long, return validity time, add some junit tests, add Builder Pattern to settings

+++ idscp2/src/main/java/de/fhg/aisec/ids/idscp2/idscp_core/finite_state_machine/StateWaitForDatAndRat.java
@@ -98,7 +98,7 @@ public class StateWaitForDatAndRat extends State {
                     LOG.debug("Verify received DAT");
                     //check if Dat is available and verify dat
                     byte[] dat = event.getIdscpMessage().getIdscpDat().getToken().toByteArray();
-                    int datValidityPeriod;
+                    long datValidityPeriod;
                     if (0 > (datValidityPeriod = dapsDriver.verifyToken(dat, null))){
                         LOG.debug("No valid remote DAT is available. Send IDSCP_CLOSE");
                         fsm.sendFromFSM(IdscpMessageFactory.getIdscpCloseMessage("No valid DAT", IDSCPv2.IdscpClose.CloseCause.NO_VALID_DAT));
