+++ idscp2/src/main/java/de/fhg/aisec/ids/idscp2/idscp_core/finite_state_machine/Timer.java
@@ -16,12 +16,12 @@ public class Timer {
         this.timeoutHandler = timeoutHandler;
     }
 
-    void resetTimeout(int delay){
+    void resetTimeout(long delay){
         cancelTimeout();
         start(delay);
     }
 
-    public void start(int delay){
+    public void start(long delay){
         mutex.lock();
         thread = new TimerThread(delay, timeoutHandler, fsmIsBusy);
         thread.start();
